사진빼고 연동 o

+++ src/main/java/com/web2/review/ReviewController.java
@@ -2,6 +2,7 @@ package com.web2.review;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.web2.global.Aop.SecureEndPoint;
 import com.web2.global.s3.S3Service;
 import com.web2.global.SessionService;
 import com.web2.restaurant.RestaurantRepository;
@@ -28,15 +29,14 @@ public class ReviewController {
     private final RestaurantRepository restaurantRepository;
     private final S3Service s3Service;
 
+    @SecureEndPoint
     @PostMapping("/reviews/new")
     public ResponseEntity<String> createReview(@RequestParam Long restaurantId,
-                                               @CookieValue(value = "SESSION_ID", required = false) String sessionId,
                                                @RequestPart("reviewDTO") String reviewDTOString,
-                                               @RequestPart(value = "image", required = false) MultipartFile image, //이미지 파일을 추가로 받음
-                                               HttpSession session) {
+                                               /*@RequestPart(value = "image", required = false) MultipartFile image,*/ //이미지 파일을 추가로 받음
+                                               HttpSession session,
+                                               @CookieValue(value = "SESSION_ID", required = false) String sessionId) {
         try {
-            sessionService.validateSession(sessionId, session);
-            sessionService.validateCsrfToken(session);
             User user = sessionService.validateUser(session);
 
             // JSON 파싱: reviewDTOString을 JSON 문자열로 받아서 ReviewDTO 객체로 변환
@@ -44,12 +44,14 @@ public class ReviewController {
             ReviewDTO reviewDTO = objectMapper.readValue(reviewDTOString, ReviewDTO.class);
 
             // 이미지 S3 업로드 로직 추가
+/*
             String imageUrl = s3Service.uploadFile(image);  // 업로드 후 이미지 URL 반환
+*/
 
             Review review = new Review(reviewDTO.message(), reviewDTO.rating(),
                     restaurantRepository.getReferenceById(restaurantId), user, reviewDTO.hashtags());
 
-            review.setImageUrl(imageUrl);  // 업로드한 이미지 URL을 리뷰에 설정
+            /*review.setImageUrl(imageUrl);  // 업로드한 이미지 URL을 리뷰에 설정*/
             reviewRepository.save(review);
 
             return ResponseEntity.ok("리뷰가 작성되었습니다.");
@@ -65,36 +67,37 @@ public class ReviewController {
         return ResponseEntity.ok(reviewResponseDTOS);
     }
 
-    @PatchMapping("/reviews/update/{id}")
-    public ResponseEntity<String> updateReview(@PathVariable Long id,
+    @SecureEndPoint
+    @PatchMapping("/reviews/update")
+    public ResponseEntity<String> updateReview(@RequestParam Long reviewId,
                                                @RequestPart("updateDTO") String updateDTOString,
-                                               @RequestPart(value = "image", required = false) MultipartFile image,
+                                               /*@RequestPart(value = "image", required = false) MultipartFile image,*/
                                                HttpSession session,
                                                @CookieValue(value = "SESSION_ID", required = false) String sessionId) throws JsonProcessingException {
-        sessionService.validateSession(sessionId, session);
-        sessionService.validateCsrfToken(session);
-
+        /*User user = sessionService.validateUser(session);*/
         ObjectMapper objectMapper = new ObjectMapper();
         ReviewUpdateDTO updateDTO = objectMapper.readValue(updateDTOString, ReviewUpdateDTO.class);
 
-        String imageUrl = s3Service.uploadFile(image);  // 업로드 후 이미지 URL 반환
+        /*String imageUrl = s3Service.uploadFile(image); */ // 업로드 후 이미지 URL 반환
 
-        Review review = reviewService.updateReview(updateDTO, id, imageUrl);
+        Review review = reviewService.updateReview(updateDTO, reviewId);
         reviewRepository.save(review);
 
         return ResponseEntity.ok("리뷰가 수정되었습니다.");
     }
 
+    @SecureEndPoint
     @DeleteMapping("/reviews/delete/{id}")
     public ResponseEntity<String> deleteReview(@PathVariable Long id,
                                                HttpSession session,
                                                @CookieValue(value = "SESSION_ID", required = false) String sessionId) {
-        sessionService.validateSession(sessionId, session);
-        sessionService.validateCsrfToken(session);
+        User user = sessionService.validateUser(session);
+
+        if (user == session.getAttribute("user")) {
+            reviewRepository.deleteById(id);
+        }
 
         //리뷰 삭제되면 S3 객체 사라지는 메서드 추가하기
-        reviewRepository.deleteById(id);
-
         return ResponseEntity.ok("리뷰가 삭제되었습니다.");
     }
